requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
Django==5.2.3
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
fonttools==4.58.5
html5lib==1.1
idna==3.10
jwcrypto==1.5.6
lxml==6.0.0
oauthlib==3.3.1
oscrypto==1.3.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.7.0
pyphen==0.17.2
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
